==> 1-Collect-Transform-OnPrem.ipynb <==
# Using the SHAP library, import the Adult Census dataset as shown following:
import shap
X, y = shap.datasets.adult()
X_display, y_display = shap.datasets.adult(display=True)
feature_names = list(X.columns)
feature_names

display(X.describe())
hist = X.hist(bins=30, sharey=True, figsize=(20, 10))

from sklearn.model_selection import train_test_split
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=1)
X_train_display = X_display.loc[X_train.index]

X_train, X_val, y_train, y_val = train_test_split(X_train, y_train, test_size=0.25, random_state=1)
X_train_display = X_display.loc[X_train.index]
X_val_display = X_display.loc[X_val.index]

import pandas as pd
train = pd.concat([pd.Series(y_train, index=X_train.index,
                             name='Income>50K', dtype=int), X_train], axis=1)
validation = pd.concat([pd.Series(y_val, index=X_val.index,
                            name='Income>50K', dtype=int), X_val], axis=1)
test = pd.concat([pd.Series(y_test, index=X_test.index,
                            name='Income>50K', dtype=int), X_test], axis=1)


train

validation

test

# Use 'csv' format to store the data
# The first column is expected to be the output column
train.to_csv('train.csv', index=False, header=False)
validation.to_csv('validation.csv', index=False, header=False)

import boto3, os

!pip install python-dotenv
%reload_ext dotenv
%dotenv

# Access and Credential details
bucket_name = os.environ['BUCKET_NAME'] # <--- Environment variable for Bucket Name
ecs_access_key_id=os.environ['ECS_ACCESS_KEY_ID']  # <--- Environment variable for ECS Access Key
ecs_secret_access_key=os.environ['ECS_SECRET_ACCESS_KEY'] # <--- Environment variable for Secret Access Key
endpoint_url=os.environ['ENDPOINT_URL'] # <--- Environment variable for Endpoint URL

s3 = boto3.resource(service_name='s3',aws_access_key_id=ecs_access_key_id,aws_secret_access_key=ecs_secret_access_key,endpoint_url=endpoint_url)


for bucket in s3.buckets.all():
  print(bucket.name)

bucket_name = os.environ['BUCKET_NAME'] # <--- Environment variable for Bucket Name

s3=boto3.resource(service_name='s3',aws_access_key_id=ecs_access_key_id,aws_secret_access_key=ecs_secret_access_key,endpoint_url=endpoint_url) 

upload_train=s3.Bucket(bucket_name).Object(os.path.join('data/train.csv')).upload_file('train.csv')
upload_validate=s3.Bucket(bucket_name).Object(os.path.join('data/validation.csv')).upload_file('validation.csv')
